requirements.txt

Django==4.1.3
django-allauth==0.51.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
Pillow==9.3.0
